Completes decentralized view
