+++ Backend/Decentralized/flask/requirements.txt
@@ -1,0 +1,2 @@
+flask
+web3
